% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_extract_meta.R
\name{extract_meta}
\alias{extract_meta}
\title{Extract metadata from images}
\usage{
extract_meta(
  ims,
  tags = c("FileName", "UserLabel", "DatetimeOriginal", "TriggerMode", "MoonPhase",
    "AmbientTemperatureFahrenheit", "AmbientTemperature", "SceneCaptureType"),
  install = F
)
}
\arguments{
\item{ims}{a vector containing file paths to images.}

\item{tags}{a vector containing the data columns you want to extract. To see the list of available tags, go to https://exiftool.org/TagNames/}

\item{install}{Default is FALSE. if TRUE, installs Exiftool}
}
\value{
A new data frame with dataset1 and dataset2 merged together.
}
\description{
Extracts metadata from images.
}
