% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_check_image.R
\name{TS_check_image}
\alias{TS_check_image}
\title{Image file check}
\usage{
TS_check_image(f)
}
\arguments{
\item{f}{a vector containing file paths to images}
}
\value{
a vector of Trues and Falses indicating if the files are image files
}
\description{
Checks if a file is an image file.
}
