% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_extract_meta.R
\name{TS_extract_meta}
\alias{TS_extract_meta}
\title{Extract metadata from images}
\usage{
TS_extract_meta(
  ims,
  tags = c("SourceFile", "UserLabel", "DatetimeOriginal", "TriggerMode", "MoonPhase",
    "AmbientTemperatureFahrenheit", "AmbientTemperature", "SceneCaptureType")
)
}
\arguments{
\item{ims}{a vector containing file paths to images.}

\item{tags}{a vector containing the data columns you want to extract. To see the list of available tags, go to https://exiftool.org/TagNames/}
}
\value{
a data frame of image metadata.
}
\description{
Extracts metadata from images.
}
