% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_launch_app.R
\name{TS_launch_app}
\alias{TS_launch_app}
\title{Launch the camTS app
Launches the app for viewing time series data and associated images.}
\usage{
TS_launch_app()
}
\value{
A Shiny window
}
\description{
Launch the camTS app
Launches the app for viewing time series data and associated images.
}
