% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_match_data.R
\name{TS_match_data}
\alias{TS_match_data}
\title{Matches observations of two datasets by datetime}
\usage{
TS_match_data(dataset1, dataset2)
}
\arguments{
\item{dataset1}{a data frame object containing a column called "Datetime". Datetime column must contain POSIXct date-time objects.}

\item{dataset2}{another data frame containing a column called "Datetime". Datetime column must contain POSIXct date-time objects. Some data from this dataset may get lost.}
}
\value{
A new data frame with dataset1 and dataset2 merged together.
}
\description{
Binds all columns from the second dataset to the first dataset by finding the closest Datetime observations.
}
